pridane testy pre Simplify
git-svn-id: svn+ssh://aiya.ms.mff.cuni.cz/var/svn/weverca/trunk@667 67bba403-d36b-46c0-a350-761a85a4c57d

+++ VSProjects/Realisation/Weverca.Analysis/MemoryAssistant.cs
@@ -757,8 +757,8 @@ namespace Weverca.Analysis
         public override void VisitAnyFloatValue(AnyFloatValue value)
         {
             containsFloat = true;
-            minFloat = Math.Min(minFloat, float.MinValue);
-            maxFloat = Math.Max(maxFloat, float.MaxValue);
+            minFloat = Math.Min(minFloat, double.MinValue);
+            maxFloat = Math.Max(maxFloat, double.MaxValue);
         }
 
         /// <inheritdoc />
